requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
pytesseract
psycopg2-binary
groq
streamlit
requests